netlink: Support accessing network namespaces.
In the future we need to access routing table entries in network
namespaces.
As netlink has no easy way to get a socket in a network namespace we
copy the logic of frr where we:
1. enter the target netns
2. open a socket
3. jump back to our original netns

the socket will keep working between these jumps.

Additionally we add logic for netlink-notifier and route-table to use
this new function.

Signed-off-by: Felix Huettner <felix.huettner@stackit.cloud>