log: Wrap log data in json key.
Previously we assumed all data in the ovsdb log is part of the data the
upper layers (e.g. raft).
In order to also support metadata in the future that just concerns the
log itself we wrap the data in a nested json structure. This allows us
to continue to pass data to the upper layers that is relevant to them
and still have metadata.

We are still able to parse the old data that is not yet wrapped in such
a nested json. With the next snapshot this will be converted.

Signed-off-by: Felix Huettner <felix.huettner@stackit.cloud>