ovsdb: Support online compactions.
Prevously compactions used snapshots in the background. While the
compaction data was prepared asynchronously the actual writing of the
data needed to happen within the main thread in order to ensure
consistency.
We can solve this differently by using multiple files within the ovsdb
log. By splitting files when the compaction thread starts we can remove
most of that dependency.

The log will create a new "curr" file and move the previous "curr" to "old".
The files "base" and "old" are read-only from then on and the compaction
thread is reponsible to write all content of them (including e.g. raft
states) to a new log file.
In the mean time "curr" can be used as normal an be written to.
At the end of the compaction we just replace "base" with the newly
written compacted file and drop "old".

This means the part of the main thread that is blocked for compaction is
now limited to writing 2 redirects, opening and renaming files. We no
longer need to write significant amounts of data in there.

The only exception of this is if compaction fails for some reason and we
need to abort to the previous state (e.g. if a disk write fails). Then
the revert will happen in the main thread to ensure consistency.

Signed-off-by: Felix Huettner <felix.huettner@stackit.cloud>