raft: Allow compaction when being leader.
Since the previous commits significantly reduced the amount of main
thread blocking during compaction we can now again allow compaction
during leadership.

As leadership changes are quite costly (especially to other applications
needing to reconnect and find the new leader) the overhead of the
compaction is generally lower then that of a leadership switch.

Signed-off-by: Felix Huettner <felix.huettner@stackit.cloud>